refactor(font-mapping): Rename folder to be less generic/more descriptive.
