+++ itest/src/FontMapper.test.ts
@@ -2,7 +2,7 @@ import { ApplicationWrapper } from "./aut/ApplicationWrapper";
 import "./expect/ElementHandleExpectations";
 import ClipboardBrowserAccessor from "./browser/ClipboardBrowserAccessor";
 import * as fs from "fs";
-import ReplaceAllPasteAction from "./actions/ReplaceAllPasteAction";
+import ReplaceAllPasteAction from "./user-interaction/ReplaceAllPasteAction";
 
 const CHARACTER_PLACEHOLDER = "{PLACE_HOLDER}";
