+++ itest/src/user-interaction/ReplaceAllPasteAction.ts
